requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphing==0.0.11
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
jax==0.4.8
joblib==1.3.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.0.4
mpmath==1.3.0
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
patsy==0.5.3
Pillow==9.5.0
platformdirs==2.5.3
plotly==5.11.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.9.3
six==1.16.0
statsmodels==0.13.5
sympy==1.12
tenacity==8.1.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
termcolor==2.2.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.16.6
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
